pyproject.toml
[tool.poetry]
name = "nonebot-plugin-calc24"
version = "0.1.0"
description = "A 24-point game plugin implemented using NoneBot."
authors = ["AwAjie <139576615+ajdgg@users.noreply.github.com>"]
license = "LICENSE"
readme = "README.md"
homepage = "https://github.com/ajdgg/nonebot-plugin-calc24"
repository = "https://github.com/ajdgg/nonebot-plugin-calc24"
documentation = "https://github.com/ajdgg/nonebot-plugin-calc24"
keywords = ["nonebot", "nonebot2"]


[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.2.0"
nonebot-adapter-onebot = { version = "^2.0.0-beta.1", optional = true }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
